TACO-653 separate tls configuration into separate artifact
