+++ build.gradle
@@ -73,11 +73,7 @@ task verifyBintrayConfig {
   }
 }
 
-task imWithStupid(type:Exec) {
-  commandLine './gradlew', '-info', 'bintrayUpload'
-}
-
-afterReleaseBuild.dependsOn imWithStupid
+afterReleaseBuild.dependsOn ':bintrayUpload'
 
 task wrapper(type: Wrapper) {
   gradleVersion = '4.7'
