+++ gradle.properties
@@ -22,7 +22,7 @@ libthrift_version=0.9.3
 zipkin_brave_brave_context_slf4j_version=5.2.0
 zipkin_reporter_zipkin_sender_okhttp3_version=2.7.7
 auto_value_version=1.5.2
-lombok_version=1.16.20
+lombok_version=1.18.2
 hamcrest_version=1.3
 metrics_version=4.0.0
