+++ settings.gradle
@@ -2,6 +2,7 @@ rootProject.name = 'xio-parent'
 // Special case as xio-core is xio
 include ':xio'
 project(':xio').projectDir = "$rootDir/xio-core" as File
+include 'xio-tls'
 include 'xio-test'
 include 'xio-zookeeper'
 include 'configuration-server'
