+++ xio-core/src/test/java/com/xjeffrose/xio/http/EdgeProxyFunctionalTest.java
@@ -6,9 +6,6 @@ import com.google.common.collect.ImmutableMap;
 import com.google.common.collect.ImmutableSet;
 import com.typesafe.config.Config;
 import com.typesafe.config.ConfigFactory;
-import com.xjeffrose.xio.SSL.MutualAuthHandler;
-import com.xjeffrose.xio.SSL.TlsAuthState;
-import com.xjeffrose.xio.SSL.TlsConfig;
 import com.xjeffrose.xio.application.Application;
 import com.xjeffrose.xio.application.ApplicationConfig;
 import com.xjeffrose.xio.application.ApplicationState;
@@ -17,6 +14,9 @@ import com.xjeffrose.xio.core.SocketAddressHelper;
 import com.xjeffrose.xio.pipeline.SmartHttpPipeline;
 import com.xjeffrose.xio.test.JulBridge;
 import com.xjeffrose.xio.test.OkHttpUnsafe;
+import com.xjeffrose.xio.tls.MutualAuthHandler;
+import com.xjeffrose.xio.tls.TlsAuthState;
+import com.xjeffrose.xio.tls.TlsConfig;
 import com.xjeffrose.xio.tracing.XioTracing;
 import io.netty.channel.ChannelHandler;
 import io.netty.channel.ChannelHandlerContext;
