+++ xio-core/src/test/java/com/xjeffrose/xio/http/ProxyWiringTest.java
@@ -4,7 +4,6 @@ import static com.xjeffrose.xio.helpers.TlsHelper.getKeyManagers;
 
 import com.google.common.collect.ImmutableMap;
 import com.typesafe.config.ConfigFactory;
-import com.xjeffrose.xio.SSL.TlsConfig;
 import com.xjeffrose.xio.application.Application;
 import com.xjeffrose.xio.bootstrap.ApplicationBootstrap;
 import com.xjeffrose.xio.helpers.ProxyPipelineRequestHandler;
@@ -12,6 +11,7 @@ import com.xjeffrose.xio.pipeline.SmartHttpPipeline;
 import com.xjeffrose.xio.pipeline.XioPipelineFragment;
 import com.xjeffrose.xio.test.JulBridge;
 import com.xjeffrose.xio.test.OkHttpUnsafe;
+import com.xjeffrose.xio.tls.TlsConfig;
 import io.netty.channel.ChannelHandler;
 import java.io.IOException;
 import java.net.InetSocketAddress;
