+++ xio-tls/build.gradle
@@ -1,0 +1,42 @@
+plugins {
+  id 'java-library'
+  id 'com.google.protobuf' version '0.8.5'
+  id 'com.github.kt3k.coveralls' version '2.8.2'
+  id 'jacoco'
+  id 'com.jfrog.bintray'
+}
+
+description = 'xio-tls'
+group = project_group
+
+bintray_package = 'xio-tls'
+publishing {
+  publications {
+    maven(MavenPublication) {
+      groupId project_group
+      artifactId bintray_package
+      version version
+      from components.java
+    }
+  }
+}
+
+bintray {
+  user = bintray_user //this usually comes from gradle.properties file in ~/.gradle
+  key = bintray_apikey //this usually comes from gradle.properties file in ~/.gradle
+  publications = ['maven'] //this refers to publishing.publications.maven up above ^^^
+  dryRun = bintray_dryrun.toBoolean() //Whether to run this as dry-run, without deploying
+  publish = true //If version should be auto published after an upload
+  pkg {
+    userOrg = bintray_subject
+    repo = bintray_repo
+    name = bintray_package
+    version {
+      name = project.version //Bintray logical version name
+    }
+    licenses = ['Apache-2.0']
+    websiteUrl = 'https://github.com/xjdr/xio'
+    issueTrackerUrl = 'https://github.com/xjdr/xio/issues'
+    vcsUrl = 'https://github.com/xjdr/xio.git'
+  }
+}
