+++ xio-tls/src/main/java/com/xjeffrose/xio/tls/X509Certificate.java
@@ -14,36 +14,19 @@
  * limitations under the License.
  *
  */
-
-package com.xjeffrose.xio.SSL;
+package com.xjeffrose.xio.tls;
 
 import java.security.PrivateKey;
+import lombok.Getter;
+import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import sun.security.x509.X509CertImpl;
 
 @Slf4j
+@RequiredArgsConstructor
+@Getter
 public final class X509Certificate {
-
   private final String fqdn;
   private final PrivateKey key;
   private final X509CertImpl cert;
-
-  public X509Certificate(String fqdn, PrivateKey key, X509CertImpl cert) {
-
-    this.fqdn = fqdn;
-    this.key = key;
-    this.cert = cert;
-  }
-
-  public String getFqdn() {
-    return fqdn;
-  }
-
-  public PrivateKey getKey() {
-    return key;
-  }
-
-  public X509CertImpl getCert() {
-    return cert;
-  }
 }
